pyproject.toml
[tool.poetry]
name = "ai-github-action"
version = "0.1.0"
description = "GitHub Action for AI-powered automation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.58.1"
requests = "^2.32.3"
PyGithub = "^2.1.1"
GitPython = "^3.1.40"
PyYAML = "^6.0.1"
pydantic = "^2.5.0"
glob2 = "^0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
mypy = "^1.0.0"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
github-agent = "src.main:main" 